exception handling and sale logic fixed

+++ src/main/java/com/ufpb/geekspace/service/AdministratorService.java
@@ -7,9 +7,12 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import com.ufpb.geekspace.exception.DataAlreadyExistsException;
+import com.ufpb.geekspace.exception.DataNotFoundException;
 import com.ufpb.geekspace.model.Administrator;
 import com.ufpb.geekspace.model.Client;
 import com.ufpb.geekspace.repository.ClientRepository;
+import com.ufpb.geekspace.util.UserUtil;
 
 @Service
 public class AdministratorService {
@@ -29,9 +32,12 @@ public class AdministratorService {
 		return aux;
 	}
 
-	public ResponseEntity<?> createAdministrator(Administrator administrator) {
+	public ResponseEntity<?> createAdministrator(Administrator administrator) throws Exception {
 		
-		Client c = new Client();
+		Client c = clientRepository.findByEmail(administrator.getEmail());
+		if(c != null)
+			throw new DataAlreadyExistsException(UserUtil.USER_ALREADY_EXISTS);
+		c = new Client();
 		c.setEmail(administrator.getEmail());
 		c.setPassword(administrator.getPassword());
 		c.setRole(administrator.getRoles());
@@ -39,7 +45,10 @@ public class AdministratorService {
 		return new ResponseEntity<Client>(c, HttpStatus.OK);
 	}
 
-	public void editAdministrator(Administrator administrator) {
+
+	public void editAdministrator(Administrator administrator) throws DataNotFoundException {
+		if(clientRepository.getOne(administrator.getId()) == null)
+			throw new DataNotFoundException(UserUtil.USER_NOT_FOUND);
 		Client c = new Client();
 		c.setId(administrator.getId());
 		c.setEmail(administrator.getEmail());
@@ -48,7 +57,9 @@ public class AdministratorService {
 		clientRepository.save(c);
 	}
 
-	public void deleteAdministrator(long administratorId) {
+	public void deleteAdministrator(long administratorId) throws DataNotFoundException {
+		if(clientRepository.getOne(administratorId) == null)
+			throw new DataNotFoundException(UserUtil.USER_NOT_FOUND);
 		clientRepository.deleteById(administratorId);
 		
 	}
